Add Redis Cluster validator
If we want to use Redis Cluster, we won't be able to use a multi-key
command with keys that hash to different slots. For instance, a simple:

    MGET foo bar

Won't work because 'foo' and 'bar' hash to different slots, and
therefore could be on different shards of the cluster.

This is a client-side validator to ensure that we can annotate existing
cross-slot commands easily, as part of the data gathering for whether or
not we use Redis Cluster.