Make `gitlab:elastic:index` enable `elasticsearch_indexing`
Before this change, the `gitlab:elastic:index` command would always fail
the on the first run, as it needed a setting to be set in the UI in the
middle of the execution.

Now this setting will be set automatically, thus making this rake task
the easiest way to populate the Elasticsearch index for a small GitLab
instance.