Add email and email_verified claims to OAuth ID tokens
This adds the "email" and "email_verified" claims to OAuth ID tokens.

This is to support OpenID Connect clients that expect the email address
in the ID token instead of the UserInfo endpoint.

Note that the claims are added to the ID token regardless of whether the
email scope is granted or not. This is currently a limitation of the
doorkeeper-openid_connect gem.