Use FLOAT_TYPE for storage limit
INT would not be enough to cover storage limits we are dealing with
since it's send in bytes. It also gets stored as bigint in the
database.

We're using FLOAT_TYPE here since BIG_INT is not exposed as a type
in our GraphQL lib at this time and it would require us to convert
it to a string since INT bigger than 32 bit are not supported:
https://spec.graphql.org/June2018/#sec-Int