Prevent autosave when reply comment via cmd+enter
Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/216631

When a user reply to a comment  using cmd + enter
and trigger another reply, the previous comment reappears.
In this commit, we fix this problem by prevent an autosave when user
reply to a comment using cmd + enter .
So when the user trigger another reply, it will be a clean slate.