Fix existing migrations for repository_storages_weighted column
For instances that upgraded to GitLab 13.1.x, it's likely that the
`application_settings.repository_storages_weighted` column did not
properly update. This happened because the schema for
`ApplicationSetting` may have been cached, so the newly-added column
wasn't visible to the model.

To ensure the column always has the right data, we make the
pre-deployment migration fill in the right data and flush the
cache. This essentially does the same thing as the post-deployment
migration, but it ensures that the application can properly run before
post-deployment migrations have completed.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/225588