Split repo move tests
The test to move from Gitaly to Gitaly Cluster wasn't running
because the Praefect scenario excludes orchestrated tests. However,
that scenario is intended to run all tests on an instance with
Praefect as storage, so it's fine to include this repo move test.

On the other hand, the repository storage scenario is like other
orchestrated tests, so the test needs the :orchestrated tag