Fix error 500s creating new projects due to empty weights
GitLab 13.1.0 added the
`application_settings.repository_storages_weighted` column, but it was
possible that the value was not seeded properly due to a stale schema
cache due to the use of ActiveRecord inside the migrations. Admins that
upgraded might see 500 errors when creating projects.

To ensure that the data is seeded properly, we surgically add a
migration that seeds the data only if it is not present. The schema
version was intentionally used to avoid backporting conflicts and to
ensure any background migrations will have the correct data when they
run. The changes were largely taken from
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/35814.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/225203