Use IP address from request on audit events
IP address captured in `current_sign_in_ip` does not refresh if the user
re-uses an existing session. It is the last IP used to sign in. IP
address from `request` object is fresher and should be use as default.

IP address in details hash needs to be extracted in the initializer to
avoid being overridden by the chaining building method (i.e. for_).