Feature flags cannot be picked after RC42
Stable branches are created 2 working days prior to the 22nd. If a
feature flag removal merge request is merged 2 working days before 
the 22nd, it will not be possible to include it in the monthly
release without delaying the whole process.

Because we deploy master changes every day, merging 3 or 4 working days
before the 22nd should be enough to get the feature in the release.