Add exclusive lock on project in ElasticCommitIndexerWorker
Indexing the same project twice at the same time causes conflict errors
in Elasticsearch and this usually leads to failures which can lead to a
project never being successfully indexed correctly. The simplest
solution is to use a lock in the worker and just fail immediately if
it's already running and let Sideikiq retries requeue the project if it
does have more updates that need indexing. When it next runs it will
check if there are new commits that need indexing.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/32648