Use the test summary endpoint
The summary endpoint is behind a feature flag.
We can use it to load the report faster, and then
load the rest of the report when the user clicks on
a specific test suite.

- Moved the dataset to be on the actual test report
element
- Changed `selectedSuite` to be `selectedSuiteIndex`
so that the suite would update after the rest of the
report was fetched.
- Changed `toggleLoading` to `setLoading` because
without the feature flag on multiple requests could be
happening and toggle loading at the same time. We
need to be able to toggle loading with the summary
endpoint feature flag on though. So being specific
about what we are setting the loading to is helpful.