Allow setting extra Sentry tags from environment
When GitLab reports an exception from Sentry, it adds tags of
'correlation_id', 'locale', and 'program'. This allows further tags to
be specified by the administrator in the GITLAB_SENTRY_EXTRA_TAGS
environment variable.

This variable must be a JSON-encoded hash of tags to send on every
Sentry event. If it fails to decode from JSON and convert to a hash, it
will simply be empty.