Change selectedSuite to use index
`selectedSuite` in the test reports store was
setting the whole suite object. This worked, but
we are moving towards not having the whole
suite object loaded at first. If we were to keep this,
then when clicking on the suite, the report wouldn't
show any test cases because they hadn't loaded yet.

Instead, we can set the index of the selectedSuite so
that when we load the whole suite we can still use
the index to get the testCases from the suite.