Fix 404 errors when "No epic" selected
If a user clicked on "No epic" in creating an issuable, a 404 would
return, and the user would lose all the information entered.  This
happened because when the button were selected, the frontend would make
the hidden element active and send the backend a "0" value. To avoid
this, we now ignore zero values.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/226999