Fix wrong value of checkbox between Rails / Vue
When a checkbox field is saved in the integration form,
in Rails (without integration_form_refactor ff) it gets saved
as "1" for true and "0" for false, while in Vue we use "true"
and "false". This caused a 1-time transition mismatch
between the values of checkboxes when the feature flag
was enabled.