Use the test report summary
The test report summary endpoint is behind a
feature flag `build_report_summary`. We can
use this endpoint to load the summary of the
test report first, and then load the rest of the
report when the user clicks on a test suite.

This `test_report_summary` in the pipeline
also allows us to render the total count in the tab
without us having to update the badge in JS.