Attempt to make signed_commits_spec.rb less flaky
The examples in this spec often attempt to click a popover before the
JavaScript handler is initialized. Since we don't have a good way to
detect when initialization has completed, we wait for all requests to
finish in an attempt to give the page more time to load before
attempting to click.

This won't guarantee that it solves the problem--we'd have to have some
indicator like a newly-added class--but hopefully this will work around
the problem for now.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/216530