Add pretty_generate coverage, skip Oj
Oj's pretty generate capabilities currently can't match the output
of the `json` gem. The `json` gem formats the JSON in a pretty
inconsistent and ugly way, but we have a number of specs that expect
it, and so I've added test coverage to ensure it matches.

Because of this, Oj usage is currently disabled for .pretty_generate.