Make sure feature flag is independent
The feature flags `vue_issuables_list` and `jira_integration`
should not be tied together, i.e. we should be able to
enable just `jira_integration` and it should work without
the need to enable the other feature flag.

This change makes sure the frontend considers this and
still renders the Jira issues list when the ff is enabled.