Use mirror when cloning to improve easy method
It is possible to download all advertised refs which gives much greater
coverage of the repository. This updates the docs to improve the first
method significantly rather than relying on the project export which is
more complex.

The first method is now more complete by including the commit map upload
step which was previously only included in the second method.