Read and parse LSIF references
In order to be able to "Find references" of a given range or
definition, we need to be able to read `references` items from
the LSIF file.

This will then be written to the JSON file for each range via
`references` property.

The property will look something like:

```
"references": [
  {
    "path": "main.go#L7"
  }
]
```

Each reference will be an object with `path` property. The `path`
property will point to the exact line where it is being used.

This is currently behind a feature flag (`code_navigation_references`)
which will set `ProcessLsifReferences` header to `true` when
enabled. If `false`, the `references` won't be read and generated.