Fix a flaky profiler spec
The check against `uniq` was meant to ensure we had multiple entries for
the same time, because the profiled code has several items that take
close to zero seconds.

However, as we're now checking against all running threads, this doesn't
really make sense: not all running threads will have that property. And
even if they did, this wasn't a great spec anyway, because it's an
unlikely failure condition.