Increase batch_size to 10000
The worker renqueue itself over and over only if at
least one registry was created or deleted by the reigstry
consistency service at least one registry class in the
next batch of replicable records. But when selective
sync is enabled, and it is iterating through a large
table most of the time, the registry consistency service
will not find an untracked/unused entry to create or
delete and will return false. This way, the worker will
not be reenqueued. Instead, we will wait for the next
one scheduled by sidekiq-cron.