Implement cronjob to adjust vulnerability statistics everyday
This cronjob will schedule background jobs for each 1000 projects to
adjust their vulnerability statistics. The reason for this approach is
if we try to adjust the vulnerability statistics of all the projects in
one background job, then in case if the cluster crushes or shuts down
with force then we can not update the statistics.