Present appropriate error for duplicate open alerts
If a triggered or acknowledged alert has a fingerprint which
matches a resolved alert, we do not allow the resolved alert
to be opened. This shows an appropriate error for the user
outlining remediation steps. This error is visible from both
the alert detail view and the alert list view.