Consolidate `trigger-build-docs` into `trigger-build`
This consolidates `trigger-build-docs` into `trigger-build` and make
the `review-docs-deploy` job wait for the downstream `gitlab-docs`
pipeline before finishing.

It was reported in Slack that the "Review "View App" button for the
docs Review App was shown even though the Review App wasn't actually
available.

That was because after triggering the downstream pipeline, we didn't
wait for it to be finished before the `review-docs-deploy` job.

Signed-off-by: Rémy Coutable <remy@rymai.me>