Limit project moved e-mails to maintainers/owners
Previously the "Project was moved" email notifications were sent every
member in the project. This served a useful purpose when we did not have
project redirects, so there was no way to discover that a project had
actually moved. However, these e-mails generate a significant amount of
noise, especially with projects with large number of team members.

To cut down on these notifications, we restrict the e-mails to
project/group maintainers and owners. Later we can introduce custom
notification settings to disable this.

Part of https://gitlab.com/gitlab-org/gitlab/-/issues/30371