Improve guidance when confirmation token expired
Previously when a User attempted to use an expired Devise confirmation
token, the error message instructed them to request a new token without
explaining how to do it. The message is only displayed on the page for
requesting a new token, but the User must infer that by the context,
which was confusing to some users. This simply adds the word "below" so
Users do not need to infer - they can follow the written instruction.