Fix package_file storage computation
When package_files are created without
an explicit size, the size is updated
on an after_save hook by copying over
the size from the file object.

This is called after the statistics
hook therefore the delta is not
added to the statistics.

This fixes that issue and updates
the statistic spec to be more robust.