Move check for data to action
To make things a little cleaner, and the component
a little more lean, I moved the check for if we have
the full report/test suite data already to the fetch
test suite action. We have easier access to the state
there, and it makes more sense to be there.