Check for temporary storage size being enabled
This adds a check for the temporary storage size end date and
exposes it to the GraphQL API.

If the date is set and in the future or the current date the
namespace will not be checked if it exceeds the limit.

To make it easier for the UI to determine if the temporary
storage is currently active or not, we expose this field in our
GraphQL API. This way the UI does not need to check and compare
the date on its own.