Remove babel-plugin-rewire dependency
This is a follow-up to [removing most no-op default exports][1]. In that
MR, the "most" was due to the fact that a couple of files were still
participating in Karma tests.

Since it turns out that no Karma tests rely on `babel-plugin-rewire`,
that could be removed, allowing the final two instances of workarounds
for it to be removed as well.

Finally, the `spyOnDependency` helper was removed, since it's unused.

[1]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/37213