Remove hyperbolic tone from SQL guideline on pluck
The statement "Never use" is already sufficiently clear and does not
need to be prefixed with another level of hyperbole. It makes me
uncomfortable when I link people to this line of the docs since it feels
like such a hostile way to give someone feedback. We can remove the
passion from these docs and still convey the same importance.