Fix dependency proxy not working with object storage
Previously the dependency proxy controller would save the blob into
object storage, but the `file_store` column was never updated. As a
result, blobs could not be downloaded because it appeared to GitLab as
though the files never made it to object storage.

Just as we do with CI job artifacts, LFS files, and project exports, we
need to update this column after the CarrierWave-mounted file is saved.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/231563