Remove .com check in over_storage_limit? method
The preference is to avoid these checks whenever possible for a
variety of reasons detailed in this issue:
https://gitlab.com/gitlab-org/gitlab/-/issues/225101.  The check wasn't
required but was put in place to prevent some additional queries when
checking plan limits that aren't necessary for self-managed instances.

Also, adjusts namespace_policy_spec so that a gitlab_subscription can
be created. It is needed for storage limit check.