Split the DeleteTagsService into two smaller services
`Projects::ContainerRepository::DeleteTagsService` acts as a
facade service that will select the appropriate sub service
according to the type of container registry used.

`Projects::ContainerRepository::Gitlab::DeleteTagsService` is
used when the Gitlab Container Registry is running. It provides a
faster execution.

`Projects::ContainerRepository::ThirdParty::DeleteTagsService` is
used when a third party Container Registry is running. It provides
a slower execution time.