Remove namespace storage limit setting
This removes the frontend and backend that is related to the
namespace storage limit setting. For the MVC we initially thought
that the limit will be instance wide, this requirement changed
though and it is now based on the plan of a namespace.

To remove the column from the db, we for now ignore the column
and will do the actual removal of the column the next release.