Prevent interrupted 2FA sign-in from signing-in incorrect user
This prevents a scenario where a user with 2FA successfully
authenticates once, don't enter their OTP code, uses the back
button in their browser, enters a random username and their
correct password and the sign-in passes as valid even though
the username was not correct. This exploit still required
knowing the user's correct password and having access to the
correct OTP code generator.