Add google tag manager
This adds the google tag manager scripts on the sign in/up and trial
sign up page. To align with our privacy policy the script only gets
activated when the `google_tag_manager_id` is set in the config on
GitLab.com and when there is no signed in user.