Use Gitaly protobuf version as DiffStats cache key
This avoids the pitfalls we saw in a mixed-deployment setup. If an
updated server adds a field such as `old_path` to the DiffStats protobuf
message definition, it will serialize this JSON to Redis. However, a
server running an older version will fetch this data and encounter an
error while trying to initialize `DiffStats`.

Most of the time this protobuf definition shouldn't change, but to be
sure we use the loaded Gem version.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/233083