Remove feature flag for using UNLINK in Redis
UNLINK was added in Redis 4. We put this usage of it behind a feature
flag in case there was an issue, and also added a fallback for instances
using Redis 3.

We no longer support Redis 3 and we haven't needed to disable the
feature flag, so we can just simplify this to always use UNLINK.