Show only root namespaces for subscriptions
We only allow subscriptions for root namespaces. However, we also
rendered sub-groups when creating a subscription.
This was not an issue so far, because we only used the new
subscription sites for sign up, therefore users were not able to
create sub-groups beforehand.

This will change in the future, where we also want to use the
new subscription page design for upgrades.